// src/app/services/room-assignation/models/assignation-room-result.ts
export class AssignationRoomResult {
  id: number;
  collaboratorId: number;
  collaborator: number;
  roomId: number;
  room: number;
  turnId: number;
  turn: number;
  modalityId: number;
  modality: number;
  branchOfficeId: number;
  branchOffice: number;
  cityId: number;
  city: number;
  userCreation: string;
  userModification: string;
  dateCreation: Date;
  dateModification: Date;
  observations: string;
  isDeleted: boolean;
}