// src/app/services/payments/models/payment-detail-result.ts
export class PaymentDetailResult {
  // Parent Data
  parentName: string;
  parentRelation: string;
  parentMaritalStatus: string;
  parentDocumentType: string;
  parentDocumentNumber: string;
  parentPhoneNumber: string;

  // Room Data
  collaboratorName: string;
  room: string;
  branchOffice: string;
  city: string;
  modality: string;
  startDateKid: Date;
  antiquity: string;
}